get all tests for client
