+++ src/main/java/kg/peaksoft/bilingualb6/dto/response/TestResponseForClient.java
@@ -17,15 +17,5 @@ public class TestResponseForClient {
     private String shortDescription;
 
     private int duration;
-}
-    private Boolean isActive;
 
-    private Integer duration;
-
-    public TestResponseForClient(Long id, String title, String shortDescription, Boolean isActive) {
-        this.id = id;
-        this.title = title;
-        this.shortDescription = shortDescription;
-        this.isActive = isActive;
-    }
 }
