+++ src/main/java/kg/peaksoft/bilingualb6/dto/response/TestResponseGetTestByIdForClient.java
@@ -22,5 +22,4 @@ public class TestResponseGetTestByIdForClient {
 
     private List<QuestionResponse> questionResponses;
 
-    private Integer duration;
 }
